requirements.txt
Flask==3.0.3
Authlib==1.3.2
python-dotenv==1.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
Werkzeug==3.0.3
tzdata==2024.1
moviepy==1.0.3
Pillow==10.4.0
pyttsx3==2.90
imageio-ffmpeg==0.4.9
